Add files via upload
Number guessing game in which the user gets to choose a range of positive numbers and then guess the number which the system has picked from it. The system will give hints regarding whether the number is lesser or greater than the user guess to help the user reach the target number quicker.